Separating food logs from foods
