+++ tests/Food/FoodLogsTest.php
@@ -1,0 +1,37 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Namelivia\Fitbit\Tests;
+
+use Carbon\Carbon;
+use Mockery;
+use Namelivia\Fitbit\Api\Fitbit;
+use Namelivia\Fitbit\Food\Foods\Logs;
+
+class FoodLogsTest extends TestCase
+{
+    private $fitbit;
+    private $logs;
+
+    public function setUp()
+    {
+        parent::setUp();
+        $this->fitbit = Mockery::mock(Fitbit::class);
+        $this->logs = new Logs($this->fitbit);
+    }
+
+    public function testGettingLogs()
+    {
+        $this->fitbit->shouldReceive('get')
+            ->once()
+            ->with('foods/log/date/2019-03-21.json')
+            ->andReturn('foodLogs');
+        $this->assertEquals(
+            'foodLogs',
+            $this->logs->get(
+                Carbon::today()
+            )
+        );
+    }
+}
